Move ability classes.
