+++ api/src/main/java/com/alibaba/nacos/api/config/ability/ClientConfigAbility.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2021 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,18 +14,15 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.api.config;
-
-import java.io.Serializable;
-import java.util.Objects;
+package com.alibaba.nacos.api.config.ability;
 
 /**
- * config abilities of nacos server.
+ * config abilities of nacos client.
  *
  * @author liuzunfei
- * @version $Id: ServerConfigAbility.java, v 0.1 2021年01月24日 00:09 AM liuzunfei Exp $
+ * @version $Id: ClientConfigAbility.java, v 0.1 2021年01月24日 00:09 AM liuzunfei Exp $
  */
-public class ServerConfigAbility implements Serializable {
+public class ClientConfigAbility {
     
     /**
      * support remote metrics get.
@@ -38,22 +35,5 @@ public class ServerConfigAbility implements Serializable {
     
     public void setSupportRemoteMetrics(boolean supportRemoteMetrics) {
         this.supportRemoteMetrics = supportRemoteMetrics;
-    }
-    
-    @Override
-    public boolean equals(Object o) {
-        if (this == o) {
-            return true;
-        }
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        ServerConfigAbility that = (ServerConfigAbility) o;
-        return supportRemoteMetrics == that.supportRemoteMetrics;
-    }
-    
-    @Override
-    public int hashCode() {
-        return Objects.hash(supportRemoteMetrics);
     }
 }
