+++ api/src/main/java/com/alibaba/nacos/api/config/ability/ServerConfigAbility.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2021 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,15 +14,18 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.api.config;
+package com.alibaba.nacos.api.config.ability;
+
+import java.io.Serializable;
+import java.util.Objects;
 
 /**
- * config abilities of nacos client.
+ * config abilities of nacos server.
  *
  * @author liuzunfei
- * @version $Id: ClientConfigAbility.java, v 0.1 2021年01月24日 00:09 AM liuzunfei Exp $
+ * @version $Id: ServerConfigAbility.java, v 0.1 2021年01月24日 00:09 AM liuzunfei Exp $
  */
-public class ClientConfigAbility {
+public class ServerConfigAbility implements Serializable {
     
     /**
      * support remote metrics get.
@@ -35,5 +38,22 @@ public class ClientConfigAbility {
     
     public void setSupportRemoteMetrics(boolean supportRemoteMetrics) {
         this.supportRemoteMetrics = supportRemoteMetrics;
+    }
+    
+    @Override
+    public boolean equals(Object o) {
+        if (this == o) {
+            return true;
+        }
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        ServerConfigAbility that = (ServerConfigAbility) o;
+        return supportRemoteMetrics == that.supportRemoteMetrics;
+    }
+    
+    @Override
+    public int hashCode() {
+        return Objects.hash(supportRemoteMetrics);
     }
 }
