+++ api/src/main/java/com/alibaba/nacos/api/remote/ability/ClientRemoteAbility.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 1999-2018 Alibaba Group Holding Ltd.
+ * Copyright 1999-2021 Alibaba Group Holding Ltd.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.alibaba.nacos.api.ability;
+package com.alibaba.nacos.api.remote.ability;
 
 /**
  * remote abilities of nacos client.
